getting rid of warnings
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@16740 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/filename.cpp
@@ -1766,7 +1766,7 @@ static void MacEnsureDefaultExtensionsLoaded()
     } ;
     // we could load the pc exchange prefs here too
 
-    for ( int i = 0 ; i < WXSIZEOF( defaults ) ; ++i )
+    for ( size_t i = 0 ; i < WXSIZEOF( defaults ) ; ++i )
     {
       gMacDefaultExtensions.Add( defaults[i] ) ;
     }
